change older implementation to use int64 type

+++ examples/HSS_Nlevel_construct.cpp
@@ -31,7 +31,7 @@ namespace Hatrix {
     ColLevelMap U;
     RowLevelMap V;
     RowColLevelMap<Matrix> D, S;
-    int N, rank, height;
+    int64_t N, rank, height;
 
   private:
