+++ examples/HSS_Nlevel_construct_miro.cpp
@@ -28,7 +28,7 @@ namespace Hatrix {
     ColLevelMap U;
     RowLevelMap V;
     RowColLevelMap<Matrix> D, S;
-    int N, rank, height;
+    int64_t N, rank, height;
 
     // Generate a row slice without the diagonal block specified by 'block'. The
     // nrows parameter determines at what level the slice is generated at. Returns
