+++ examples/UMV_weak_2level.cpp
@@ -19,7 +19,7 @@ namespace Hatrix {
     ColLevelMap U;
     RowLevelMap V;
     RowColLevelMap<Matrix> D, S;
-    int N, rank, height;
+    int64_t N, rank, height;
 
   private:
